Update frontend to use custom type-safe generated hooks rather than useQuery or useMutation
